+++ docs/Changelog.md
@@ -34,6 +34,7 @@
 * [#4299](https://github.com/TouK/nussknacker/pull/4299)[4322](https://github.com/TouK/nussknacker/pull/4322) `StateStatus` is identified by its name. 
   `ProcessState` serialization uses this name as serialized state value. For compatibility reasons, it is still represented as a nested object with one `name` field.
 * [#4312](https://github.com/TouK/nussknacker/pull/4312) Fix for losing unsaved changes in designer after cancel/deploy
+* [#4331](https://github.com/TouK/nussknacker/pull/4331) Improvements: Processes API for archived and subprocss return data with no state info
 
 1.9.1 (24 Apr 2023)
 ------------------------